added cpp_int mpz

+++ include/PricerSolverArcTimeDP.hpp
@@ -97,16 +97,16 @@ class PricerSolverArcTimeDp : public PricerSolverBase {
     OptimalSolution<double> pricing_algorithm(double* _pi) override;
     OptimalSolution<double> farkas_pricing(double* pi) override;
 
-    void   iterate_zdd() override;
-    void   create_dot_zdd(const char* name) override;
-    void   print_number_nodes_edges() override;
-    size_t get_num_remove_nodes() override;
-    size_t get_num_remove_edges() override;
-    size_t get_nb_edges() override;
-    size_t get_nb_vertices() override;
-    int    get_num_layers() override;
-    size_t print_num_paths() override;
-    bool   check_schedule_set(const std::vector<Job*>& set) override;
+    void    iterate_zdd() override;
+    void    create_dot_zdd(const char* name) override;
+    void    print_number_nodes_edges() override;
+    size_t  get_num_remove_nodes() override;
+    size_t  get_num_remove_edges() override;
+    size_t  get_nb_edges() override;
+    size_t  get_nb_vertices() override;
+    int     get_num_layers() override;
+    cpp_int print_num_paths() override;
+    bool    check_schedule_set(const std::vector<Job*>& set) override;
 
     void forward_evaluator(double* pi);
     void backward_evaluator(double* _pi);
