+++ include/PricerSolverSimpleDP.hpp
@@ -66,15 +66,15 @@ class PricerSolverSimpleDp : public PricerSolverBase {
         override;
 
     // void   add_constraint(Job* job, GPtrArray* list, int order) override;
-    void   iterate_zdd() override;
-    void   create_dot_zdd(const char* name) override;
-    void   print_number_nodes_edges() override;
-    size_t get_num_remove_nodes() override;
-    size_t get_num_remove_edges() override;
-    size_t get_nb_edges() override;
-    size_t get_nb_vertices() override;
-    int    get_num_layers() override;
-    size_t print_num_paths() override;
+    void    iterate_zdd() override;
+    void    create_dot_zdd(const char* name) override;
+    void    print_number_nodes_edges() override;
+    size_t  get_num_remove_nodes() override;
+    size_t  get_num_remove_edges() override;
+    size_t  get_nb_edges() override;
+    size_t  get_nb_vertices() override;
+    int     get_num_layers() override;
+    cpp_int print_num_paths() override;
 
     bool check_schedule_set(const std::vector<Job*>& set) override;
