+++ include/PricerSolverZdd.hpp
@@ -21,8 +21,8 @@
 // #include "ZddNode.hpp"
 // #include "wctprivate.h"
 // #include <stddef.h>                              // for size_t
-#include <stddef.h>                              // for size_t
 #include <NodeBddStructure.hpp>                  // for DdStructure
+#include <cstddef>                               // for size_t
 #include <memory>                                // for unique_ptr, shared_ptr
 #include <range/v3/iterator/basic_iterator.hpp>  // for operator!=
 #include <range/v3/view/filter.hpp>              // for filter_view
@@ -85,17 +85,17 @@ class PricerSolverZdd : public PricerSolverBase {
         const double*                                    columns,
         const std::vector<std::shared_ptr<ScheduleSet>>& schedule_sets)
         override;
-    bool   check_schedule_set(const std::vector<Job*>& set) override;
-    void   iterate_zdd() override;
-    void   create_dot_zdd(const char* name) override;
-    void   print_number_nodes_edges() override;
-    size_t get_num_remove_nodes() override;
-    size_t get_num_remove_edges() override;
-    size_t get_nb_edges() override;
-    size_t get_nb_vertices() override;
-    int    get_num_layers() override;
-    size_t print_num_paths() override;
-    void   remove_layers_init();
+    bool    check_schedule_set(const std::vector<Job*>& set) override;
+    void    iterate_zdd() override;
+    void    create_dot_zdd(const char* name) override;
+    void    print_number_nodes_edges() override;
+    size_t  get_num_remove_nodes() override;
+    size_t  get_num_remove_edges() override;
+    size_t  get_nb_edges() override;
+    size_t  get_nb_vertices() override;
+    int     get_num_layers() override;
+    cpp_int print_num_paths() override;
+    void    remove_layers_init();
 
     OptimalSolution<double> farkas_pricing(double* pi) override;
